+++ src/main/java/com/ovtrip/user/model/UserEntity.java
@@ -5,14 +5,14 @@ import lombok.Data;
 
 @Data
 public class UserEntity {
-    private int userId;
+    private Long userId;
     private String userEmail;
     private String userPassword;
     private String userNickname;
     private String socialType;
 
     @Builder
-    public UserEntity(int userId, String userEmail, String userPassword, String userNickname, String socialType) {
+    public UserEntity(Long userId, String userEmail, String userPassword, String userNickname, String socialType) {
         this.userId = userId;
         this.userEmail = userEmail;
         this.userPassword = userPassword;
